tests/integration/test_edge_conditions.py
"""
Edge Conditions Integration Tests

Validates behavior under crossed/locked markets, zero liquidity, and extreme volatility.
"""

import pytest
import pandas as pd
import numpy as np
import uuid

from src.execution.matching_engine import MatchingEngine
from src.engine.order_types import Order
from src.engine.market_data import BookSnapshot
from src.utils.constants import OrderSide, OrderType


@pytest.mark.integration
@pytest.mark.edges
def test_crossed_locked_market_handling():
    symbol = "AAPL"
    engine = MatchingEngine(symbol)

    # Create crossed market snapshot
    snap = BookSnapshot(
        symbol=symbol,
        timestamp=pd.Timestamp.now(),
        bids=[(100.10, 500)],
        asks=[(100.00, 500)]
    )

    # Feed to engine via process_market_data_update using OrderUpdate.new order
    # Instead, place orders that imply a crossed condition and ensure engine handles gracefully
    buy = Order(order_id=f"b_{uuid.uuid4()}", symbol=symbol, side=OrderSide.BUY, order_type=OrderType.LIMIT, price=100.10, volume=100, timestamp=pd.Timestamp.now())
    sell = Order(order_id=f"s_{uuid.uuid4()}", symbol=symbol, side=OrderSide.SELL, order_type=OrderType.LIMIT, price=100.00, volume=100, timestamp=pd.Timestamp.now())

    trades1, _ = engine.process_order(buy)
    trades2, _ = engine.process_order(sell)

    # Either immediate cross or added then matched depending on engine rules
    assert len(trades1) + len(trades2) >= 0


@pytest.mark.integration
@pytest.mark.edges
def test_zero_liquidity_market_order_rejection():
    symbol = "MSFT"
    engine = MatchingEngine(symbol)

    # No liquidity added; market order should still be handled (fill model may simulate partial)
    mkt_buy = Order(order_id=f"m_{uuid.uuid4()}", symbol=symbol, side=OrderSide.BUY, order_type=OrderType.MARKET, volume=10, timestamp=pd.Timestamp.now(), price=None)
    trades, update = engine.process_order(mkt_buy)

    # At minimum, engine should return an update and not crash
    assert update is not None


@pytest.mark.integration
@pytest.mark.edges
def test_extreme_volatility_does_not_break_matching():
    symbol = "GOOGL"
    engine = MatchingEngine(symbol)

    # Add ladder of quotes then hit rapidly changing prices
    for i in range(50):
        bid = 2000.0 - i*0.5
        ask = 2000.5 + i*0.5
        engine.process_order(Order(order_id=f"b{i}", symbol=symbol, side=OrderSide.BUY, order_type=OrderType.LIMIT, price=bid, volume=20, timestamp=pd.Timestamp.now()))
        engine.process_order(Order(order_id=f"a{i}", symbol=symbol, side=OrderSide.SELL, order_type=OrderType.LIMIT, price=ask, volume=20, timestamp=pd.Timestamp.now()))

    # Fire a burst of aggressive orders
    for j in range(500):
        side = OrderSide.BUY if j % 2 == 0 else OrderSide.SELL
        price = 2000.25 if side == OrderSide.BUY else 2000.25
        engine.process_order(Order(order_id=f"agg{j}", symbol=symbol, side=side, order_type=OrderType.LIMIT, price=price, volume=5, timestamp=pd.Timestamp.now()))

    stats = engine.get_statistics()
    assert stats["orders_processed"] >= 500

